+++ src/main/java/com/eduflex/eduflexbackend/controller/FacultyLeaveController.java
@@ -11,7 +11,9 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.eduflex.eduflexbackend.exception.EduflexDataNotFoundException;
 import com.eduflex.eduflexbackend.model.FacultyLeave;
+import com.eduflex.eduflexbackend.repository.FacultyLeaveRepository;
 import com.eduflex.eduflexbackend.service.FacultyLeaveService;
 
 @CrossOrigin
@@ -19,32 +21,54 @@ import com.eduflex.eduflexbackend.service.FacultyLeaveService;
 public class FacultyLeaveController {
 	@Autowired
 	FacultyLeaveService facultyLeaveService;
+	
+	@Autowired
+	FacultyLeaveRepository facultyLeaveRepository;
 
 	@GetMapping("/facultyLeaves")
 	public List<FacultyLeave> getAllFacultyLeaves() {
+		if(facultyLeaveRepository.findAll().size() == 0) {
+			throw new EduflexDataNotFoundException("FacultyLeave Not Found, Add facultyLeave first");
+		}
 		return facultyLeaveService.getAllFacultyLeaves();
 	}
-	//newAdded
+	
     @GetMapping("/faculty/{facultyId}/facultyLeaves")
     public List<FacultyLeave> getFacultyLeavesByFacultyId(@PathVariable int facultyId) {
+    	if(facultyLeaveRepository.findAll().size() == 0) {
+			throw new EduflexDataNotFoundException("facultyLeave with facultyId "+facultyId+" Not Found, Add facultyLeave first");
+		}
         return facultyLeaveService.getFacultyLeavesByFacultyId(facultyId);
     }
 	
     @GetMapping("/faculty/{facultyId}/facultyLeave/{facultyLeaveId}")
 	public FacultyLeave getFacultyLeaveByFacultyLeaveId(@PathVariable int facultyId,@PathVariable int facultyLeaveId) {
+    	if(!facultyLeaveRepository.existsById(facultyLeaveId)) {
+    		throw new EduflexDataNotFoundException("Can't find facultyLeave with facultyLeaveId: "+facultyLeaveId+", please insert first");
+    	}
         return facultyLeaveService.getFacultyLeaveByFacultyLeaveId(facultyId, facultyLeaveId);
     }
 	
 	@DeleteMapping("/faculty/{facultyId}/facultyLeave/{facultyLeaveId}")
 	public void deleteFacultyLeave(@PathVariable int facultyId,@PathVariable int facultyLeaveId) {
+		if(!facultyLeaveRepository.existsById(facultyId)) {
+    		throw new EduflexDataNotFoundException("Can't delete facultyLeave, facultyId: "+facultyId+", not have FacultyLeave in database");
+    	}
 		facultyLeaveService.deleteFacultyLeave(facultyId,facultyLeaveId);
 	}
 	
 	@PutMapping("/faculty/{facultyId}/facultyLeave")
 	FacultyLeave addFacultyLeaveToFaculty(@PathVariable int facultyId,@RequestBody  FacultyLeave facultyLeave) {
+		if(!facultyLeaveRepository.existsById(facultyId)) {
+    		throw new EduflexDataNotFoundException("Can't add studentLeave to facultyId: "+facultyId+", because it not persent in database");
+    	}
 		  return facultyLeaveService.addFacultyLeaveToFaculty(facultyId, facultyLeave);}
 	
 	@GetMapping("/faculty/facultyLeaves/{facultyLeaveId}")
-	public void deleteFacultyLeaveByFacultyLeaveID(@PathVariable int facultyLeaveId)
-	{ facultyLeaveService.deleteFacultyLeaveByFacultyLeaveID(facultyLeaveId);}
+	public void deleteFacultyLeaveByFacultyLeaveID(@PathVariable int facultyLeaveId) {
+		if(!facultyLeaveRepository.existsById(facultyLeaveId)) {
+    		throw new EduflexDataNotFoundException("Can't find facultyLeave with facultyLeaveId: "+facultyLeaveId+", please insert first");
+    	}
+		facultyLeaveService.deleteFacultyLeaveByFacultyLeaveID(facultyLeaveId);
+		}
 }
